first pass at all new grammars. I eppy, will work on sprint for these grammars tmmrw

+++ hdr/helper/alloc.h
@@ -2,6 +2,6 @@
 #define ALLOC_H
 #include <stdlib.h>
 
-void *alloc(size_t size);
+void *alloc(int size);
 
 #endif
