+++ src/helper/alloc.c
@@ -1,7 +1,7 @@
 #include "alloc.h"
 #include <string.h>
 
-void *alloc(size_t size) {
+void *alloc(int size) {
   void *ret = malloc(size);
   memset(ret, 0, size);
   return ret;
